Change cart display logic
Until previous versions, this component only was displayed when the cart
had items on it, and a blue circle in the left bottom corner with a
number in its center showed the total quantity of them.

Now, the cart icon will be displayed all the time and will redirect to
the cart view. The total quantity of items into the cart is going to be
displayed only when it has at least one item inside, using the same blue
circle just mentioned before.