Add dynamic behavior to component views
The behavior of the container components on failure or slow loading was
not defined. With this modification to code, the app will display a
loading animation until all items be fetched from the database. Also, if
the query to database results empty (items not found), another view will
be displayed, explaining the absence of the expected items.