Delete not functional components from NavBar
Some visual components were added at the beginning of the development of
this project trying to mimic the frontend of the most known ecommerce
platforms.

However, neither of those components had an implemented functionality,
nor the complexity and time constraints of this project will allow the
development of those features in the near future.