Add confirmation to required fields
It's usual to require users to reinsert some info in certain forms to
assure that the data inserted is correct.

Here, as the app doesn't have authentication, to mimic that behavior,
the requested field will be the email. Although the logic has been
designed to support any number of confirmation fields.