Correct confirmation field logic
To check if the confirmation field is equal to its matching field, the
value of both fields has to be compared on each change of any of this
input fields.

To perform this action correctly a state is needed to hold the value
of the confirmation fields on every render and the useEffect hook will
do the checking on each change of the related fields.